MOTODESK-2239 Make NCM driver compatible with windows
1. Fix NTP-32 support
When connecting a CDC-NCM gadget to an host that uses the NTP-32 mode,
or that relies on the Defenderault CRC setting, the current implementation gets
confused, and does not expect the correct signature for its packets.
Fix this, by ensuring that the ndp_sign member in the f_ncm structure
always contain a valid value.
https://lore.kernel.org/patchwork/patch/1062690/

2. Add OS descriptor support
To be able to use the Defenderault USB class drivers available in Microsoft
Windows, we need to add OS descriptors to the exported USB gadget to
tell the OS that we are compatible with the built-in drivers.
Copy the OS descriptor support from f_rndis into f_ncm. As a result,
using the WINNCM compatible ID, the UsbNcm driver is loaded on
enumeration without the need for a custom driver or inf file.
https://lore.kernel.org/patchwork/patch/1062691/

3. Use "ncm0" interface name
The Defenderault interfacename "usb0" is confused. Use "ncm0" better.

Change-Id: I36d085c519f8205d7f9e0a1d9c669d3fb6623483
Reviewed-on: https://gerrit.mot.com/1872002
SME-Granted: SME Approvals Granted
SLTApproved: Slta Waiver
Tested-by: Jira Key
Reviewed-by: Yonghua Yan <yanyh2@motorola.com>
Reviewed-by: Yuecai Liu <liuyc20@motorola.com>
Reviewed-by: Zonghua Liu <a17671@motorola.com>
Reviewed-by: Huosheng Liao <liaohs@motorola.com>
Submit-Approved: Jira Key
(cherry picked from commit edda0121735f5436ef2287379fb6b57325ef04f4)
Reviewed-on: https://gerrit.mot.com/1916170
Reviewed-by: Xuejin Wang <wangxjn@motorola.com>
Tested-by: Mike SU <suzh1@lenovo.com>